refactor: User와 관련된 아키텍처에 UserDto로 대체
